Move exception classes to a separate file
