+++ modules/utils.py
@@ -7,3 +7,19 @@ def does_file_exist(filename):
 
     else:
         return False
+
+
+class WarnException(Exception):
+    def __init__(self, msg):
+        self.msg = msg
+
+    def __str__(self):
+        return self.msg
+
+
+class CriticalException(Exception):
+    def __init__(self, msg):
+        self.msg = msg
+
+    def __str__(self):
+        return self.msg
